app/core/config.py
from pydantic_settings import BaseSettings
from dotenv import load_dotenv

# Load from .env
load_dotenv()

class Settings(BaseSettings):
    DATABASE_URL: str
    SECRET_KEY: str
    DEBUG: bool = False

    class Config:
        env_file = ".env"
        env_file_encoding = "utf-8"

settings = Settings()
